+++ internal/langserver/handlers/command/handler.go
@@ -1,7 +1,10 @@
 package command
 
-import "context"
+import (
+	"context"
+	"encoding/json"
+)
 
 type CommandHandler interface {
-	Handle(ctx context.Context, params CommandArgs) (interface{}, error)
+	Handle(ctx context.Context, params []json.RawMessage) (interface{}, error)
 }