+++ internal/lsp/token_encoder.go
@@ -105,10 +105,15 @@ func (te *TokenEncoder) encodeTokenOfIndex(i int) []uint32 {
 		deltaStartChar := token.Range.Start.Column - 1 - previousStartChar
 
 		data = append(data, []uint32{
+			// #nosec G115
 			uint32(deltaLine),
+			// #nosec G115
 			uint32(deltaStartChar),
+			// #nosec G115
 			uint32(tokenLength),
+			// #nosec G115
 			uint32(tokenTypeIdx),
+			// #nosec G115
 			uint32(modifierBitMask),
 		}...)
 	} else {
@@ -129,10 +134,15 @@ func (te *TokenEncoder) encodeTokenOfIndex(i int) []uint32 {
 			}
 
 			data = append(data, []uint32{
+				// #nosec G115
 				uint32(deltaLine),
+				// #nosec G115
 				uint32(deltaStartChar),
+				// #nosec G115
 				uint32(length),
+				// #nosec G115
 				uint32(tokenTypeIdx),
+				// #nosec G115
 				uint32(modifierBitMask),
 			}...)